Move non-domain related components to ZFX module
