+++ module/Accounting/src/Accounting/Controller/StatementsController.php
@@ -2,11 +2,11 @@
 namespace Accounting\Controller;
 
 use Zend\Permissions\Acl\Acl;
-use Application\Controller\AbstractHATEOASRestfulController;
+use ZFX\Rest\Controller\HATEOASRestfulController;
 use Accounting\Service\AccountService;
 use Accounting\View\StatementJsonModel;
 
-class StatementsController extends AbstractHATEOASRestfulController
+class StatementsController extends HATEOASRestfulController
 {
 	protected static $collectionOptions = ['GET'];
 	protected static $resourceOptions = ['GET'];
