+++ module/Application/src/Application/Controller/MembershipsController.php
@@ -8,10 +8,10 @@ use Zend\Filter\StripNewlines;
 use Zend\Filter\StripTags;
 use People\Organization;
 use People\Service\OrganizationService;
-use Application\Controller\AbstractHATEOASRestfulController;
+use ZFX\Rest\Controller\HATEOASRestfulController;
 use Application\View\OrganizationMembershipJsonModel;
 
-class MembershipsController extends AbstractHATEOASRestfulController
+class MembershipsController extends HATEOASRestfulController
 {
 	protected static $collectionOptions = array('GET', 'POST');
 	protected static $resourceOptions = array('DELETE', 'POST', 'GET', 'PUT');
